added init

+++ src/main/java/donkey/koin/transactions/transaction/TransactionDetails.java
@@ -11,11 +11,11 @@ public class TransactionDetails {
     private final String username;
     private final Instant transactionTime;
     private final Double moneyAmount;
-    private final int lastKoinValue;
+    private final double lastKoinValue;
 
     public TransactionDetails(@JsonProperty("moneyAmount") Double moneyAmount,
                               @JsonProperty("username") String username,
-                              @JsonProperty("lastKoinValue") int lastKoinValue,
+                              @JsonProperty("lastKoinValue") double lastKoinValue,
                               @JsonProperty("transactionTime") Instant transactionTime) {
         this.moneyAmount = moneyAmount;
         this.username = username;
